// src/app/models/response/instructor/getInstructor-response.ts
export interface IGetInstructorResponseModel {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  companyName: string;
  nationalIdentity: number;
  dateOfBirth: string;
  image:string;
  title:string;
  instagramadress:string;
  twitteradress:string;
  githubadress:string



}